P# On branch full-sim

+++ source/persistency/hdf5_functions.h
@@ -13,21 +13,21 @@
    } run_info_t; 
 
   typedef struct{
-    int evt_number;
-    unsigned int last_sns_data;
-    unsigned int last_sns_tof;
-    unsigned int last_hit;
-    unsigned int last_particle;
+    int32_t evt_number;
+    uint64_t last_sns_data;
+    uint64_t last_sns_tof;
+    uint64_t last_hit;
+    uint64_t last_particle;
   } evt_extent_t;
 
   typedef struct{
-    int evt_number;
+    int32_t evt_number;
     float evt_energy;
   } evt_t;
 
   typedef struct{
     unsigned int sensor_id;
-    unsigned int time_bin;
+    uint64_t time_bin;
     unsigned int charge;
   } sns_data_t;
